+++ src/orchestration/jepx_pipeline.py
@@ -21,10 +21,10 @@ import polars as pl
 
 from catalog.manage_iceberg import get_catalog, provision_table
 from common.storage_client import RustFSClient
-from pipeline.ingestion.ingest_jepx import ingest_jepx_spot_summary
-from pipeline.jepx.common import resolve_target_at
+from jepx.common import resolve_target_at
+from pipeline.bronze.ingest_jepx import ingest_jepx_spot_summary
 from pipeline.raw.jepx_to_rustfs import scrape_jepx_to_rustfs
-from pipeline.scraper.module.jepx import JEPXSpotSummaryScraper
+from scraper.module.jepx import JEPXSpotSummaryScraper
 
 logger = logging.getLogger(__name__)
