+++ src/pipeline/bronze/ingest_jepx.py
@@ -11,7 +11,7 @@ sys.path.append(str(Path(__file__).resolve().parent.parent.parent))
 
 from catalog.manage_iceberg import get_catalog
 from common.storage_client import RustFSClient
-from pipeline.jepx.common import (
+from jepx.common import (
     resolve_spot_summary_file_name,
     resolve_spot_summary_object_key,
     resolve_target_at,
