+++ src/scraper/main.py
@@ -2,9 +2,9 @@ import argparse
 import logging
 
 from common.storage_client import RustFSClient
-from pipeline.jepx.common import resolve_target_at
+from jepx.common import resolve_target_at
 from pipeline.raw.jepx_to_rustfs import scrape_jepx_to_rustfs
-from pipeline.scraper.module.jepx import JEPXSpotSummaryScraper
+from scraper.module.jepx import JEPXSpotSummaryScraper
 
 logging.basicConfig(
     level=logging.INFO, format="%(asctime)s | %(levelname)s | %(message)s"
